Revert "."
This reverts commit 176175d108f28dcd91d9c8b19aba54f243aad2dd.

+++ HabitGame/Assets/Scripts/PvP/BattleUnit.cs
@@ -5,11 +5,11 @@ using UnityEngine;
 public class BattleUnit
 {
     public string name;
-    public int maxHp;
-    public int hp;
-    public float atk;
-    public float def;
-    public float spd;
+    public float maxHp;
+    public float hp;
+    public int atk;
+    public int def;
+    public int spd;
     public float crtk;    // 0~1
 
     public AttributeType attribute;
